refactor: View 가독성/재사용성 측면 리팩토링
