+++ docs/REQUIREMENTS.md
@@ -167,7 +167,7 @@
 1. 상수 처리
    [X] Model 내부로 상수 위치 변경
    [X] 상수가 클래스와 결합해야하면 `static` 키워드 사용
-   [ ] 클래스 내부 상태 은닉 필요하면 `private` 접근 제어자 사용
+   [X] 클래스 내부 상태 은닉 필요하면 `private` 접근 제어자 사용
 
 2. 요구사항 변경 가능지점 체크하여 확장성 고려해 코드 작성
    [ ] 자동차 전진 조건이 변경되는 경우
