--- src/utils/utils.js
@@ -1,3 +1,0 @@
-export function getRandomIntRangeOf(min, max) {
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
